more docs for transaction module
